pyproject.toml
[project]
name = "llm-code"
version = "0.9.0"
description = "An OpenAI LLM based CLI coding assistant."
readme = "README.md"
license = "MIT"
authors = [{ name = "Rushabh Doshi", email = "radoshi+pypi@gmail.com" }]
keywords = ["openai", "llm", "cli", "coding", "assistant"]
classifiers = [
  "Topic :: Software Development",
  "Topic :: Software Development :: Code Generators",
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
]
requires-python = ">=3.14.3,<3.15"
dependencies = [
  "click==8.3.1",
  "pydantic==2.12.5",
  "pydantic-ai-slim[anthropic,google,openai]==1.71.0",
  "pydantic-settings==2.13.1",
  "pyperclip==1.11.0",
  "pyyaml==6.0.3",
  "rich==14.3.3",
  "sqlalchemy==2.0.48",
  "textual[syntax]==8.1.1",
]

[project.urls]
Homepage = "https://github.com/radoshi/llm-code"
Repository = "https://github.com/radoshi/llm-code"

[project.scripts]
llm-code = "llm_code.llm_code:main"
llm_code = "llm_code.llm_code:main"

[dependency-groups]
dev = [
  "pytest==9.0.2",
  "pytest-cov==7.1.0",
  "pytest-watch==4.2.0",
  "ruff==0.15.7",
  "textual-dev==1.8.0",
  "ty==0.0.25",
]

[build-system]
requires = ["hatchling==1.29.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_code"]

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
  "/README.md",
  "/LICENSE",
]

[tool.ruff]
target-version = "py314"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
